Start working on the compose page and logic to make the system integrate tighter with compose calls instead of always using docker calls
Adjust the setFile() for the migration check
Check for existing file before calling unlink in the deleteFile()
Add symfony yaml to parse compose files

+++ root/app/www/public/functions/files.php
@@ -44,6 +44,10 @@ function setFile($file, $contents)
 
 function deleteFile($file)
 {
+    if (!file_exists($file)) {
+        return;
+    }
+
     logger(SYSTEM_LOG, 'deleteFile() ' . $file);
 
     unlink($file);
